M-cycle CPU rewrite + PPU fixes: tile bit order, signed tile addressing, 60fps
- M-cycle CPU rewrite: all 256 instruction handlers rewritten with
  explicit M-cycle stepping via stepDevices(4). Removed deferred write
  system (schedWrite/execPend). All Blargg tests pass (cpu_instrs,
  instr_timing, halt_bug, oam_bug, mem_timing).
- Fixed tile pixel bit order: Game Boy tile format stores bit 7 as
  leftmost pixel, but decodeTileRow and renderSprites read bit 0 as
  leftmost. Fixed both to use (7-i) bit indexing.
- Fixed signed tile addressing base: 0x8800 -> 0x9000 (tile 0 in
  signed mode should be at 0x9000, not 0x8800).
- Changed emulation loop from 30fps to 60fps for full-speed audio.
- Added StepDevices to MMU interface, implemented in MemoryBus.
- Decoupled PPU register reads/writes via CatchUpPPU for T-cycle
  accuracy.
- Converted timer cycle counters from uint16 to uint64.
- Updated all tests for correct PPU LCD enable and 2-frame stepping.

+++ internal/gb/timer.go
@@ -21,11 +21,11 @@ type Timer struct {
 	TMA  byte // 0xFF06
 	TAC  byte // 0xFF07
 
-	// Internal 16-bit cycle counters.
+	// Internal 64-bit cycle counters (must not overflow — a frame is ~70k cycles).
 	// divCycles counts toward DIV increments (free-running, every 256 cycles).
 	// timaCycles counts toward TIMA increments (only when TAC bit 2 is set).
-	divCycles  uint16
-	timaCycles uint16
+	divCycles  uint64
+	timaCycles uint64
 
 	// mmu for requesting interrupts (set on TIMA overflow)
 	mmu MMU
@@ -48,7 +48,7 @@ func (t *Timer) Step(cycles int) {
 	}
 
 	// --- DIV: free-running, always increments ---
-	t.divCycles += uint16(cycles)
+	t.divCycles += uint64(cycles)
 	for t.divCycles >= 256 {
 		t.divCycles -= 256
 		t.DIV++ // wraps naturally
@@ -59,8 +59,8 @@ func (t *Timer) Step(cycles int) {
 		return
 	}
 
-	threshold := t.timaThreshold()
-	t.timaCycles += uint16(cycles)
+	threshold := uint64(t.timaThreshold())
+	t.timaCycles += uint64(cycles)
 
 	for t.timaCycles >= threshold {
 		t.timaCycles -= threshold
